Kernel: Split the DevPtsFS files into smaller components
