+++ Kernel/TTY/SlavePTY.cpp
@@ -6,7 +6,6 @@
 
 #include <AK/Singleton.h>
 #include <Kernel/Debug.h>
-#include <Kernel/FileSystem/DevPtsFS.h>
 #include <Kernel/Process.h>
 #include <Kernel/TTY/MasterPTY.h>
 #include <Kernel/TTY/SlavePTY.h>
